# Memory issue for SGD and sparse grad