Fixes #3129 .

Previously we use buffer with same type as grad. For sparse grad, the buffer grows continuously as we never call coalesce. Since the buffer will likely become dense anyways, this PR changes it so that buffer are init'd as same type of param.

Performance on script in #3219 
Before change: GPU memory usage fluctuates greatly, with max at 10800MB. Decaying speed from 110it/s to 44it/s in 9 epochs.
Adding coalesce in each call: Constant GPU memory usage 1231MB. Constant speed at 150it/s.
After change: Constant GPU memory usage 1197MB. Constant speed at 210it/s.