@ezyang 2017-10-16
Very nice!