@soumith 2017-10-16
for those who come to this PR after, the changes are equivalent to before (but not exactly the same)