@seba-1511 2018-07-22
I think this PR introduced a superfluous `.mul_(momentum)` on line 90. (line 98 in current version of `SGD`)

Since `buf` is initialized to be zero, multiplying it by `momentum` should be a no-op and wastes cycles. Or am I missing something ?