@backyes 2018-11-26
How this sparse gradients optimization works with distributed training?  Sparse gradient can be used to reduce gradient communication between parameters server. 
I do not know if the sparse gradient optimization can works with distributed p.s.  ?  if will, this will benefits large sparse embedding training.

Hope some comments on this, thanks.